requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
flask-cors
pandas
asyncio
httpx